Backend/test_innovate/requirements.txt
Django==6.0.3
djangorestframework==3.16.1
django-cors-headers==4.7.0
djangorestframework-simplejwt==5.5.1
